backend/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_bcrypt
psycopg2-binary
bcrypt
python-dotenv
requests
